requirements.txt
Django==4.2.1
djangorestframework==3.14.0
requests==2.28.1
python-dotenv==0.21.1
django-cors-headers==3.13.0
psycopg2-binary==2.9.5
openai==0.27.7
drf-yasg==1.21.4
